Removed Legend from Table file, removed ternary semll
